@@ -1,109 +1,107 @@
 {
-  "name": "novelscraper",
-  "version": "1.0.8",
-  "description": "Educational app on how to web scrape novels from pirate sites.",
-  "homepage": "https://github.com/dr-nyt/NovelScraper",
-  "author": {
-    "name": "dr_nyt",
-    "email": "shehryar.dev@gmail.com"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dr-nyt/NovelScraper.git"
-  },
-  "keywords": [
-    "webscraper",
-    "novel",
-    "downloader"
-  ],
-  "main": "main.js",
-  "private": true,
-  "scripts": {
-    "postinstall": "electron-builder install-app-deps",
-    "ng": "ng",
-    "start": "npm-run-all -p electron:serve ng:serve",
-    "build": "npm run electron:serve-tsc && ng build",
-    "build:dev": "npm run build -- -c dev",
-    "build:prod": "npm run build -- -c production",
-    "ng:serve": "ng serve",
-    "ng:serve:web": "ng serve -c web -o",
-    "electron:serve-tsc": "tsc -p tsconfig-serve.json",
-    "electron:serve": "wait-on http-get://localhost:4200/ && npm run electron:serve-tsc && electron . --serve",
-    "electron:local": "npm run build:prod && electron .",
-    "electron:linux": "npm run build:prod && electron-builder build --linux",
-    "electron:windows": "npm run build:prod && electron-builder build --windows",
-    "deploy:window": "npm run build:prod && electron-builder build --windows --publish always",
-    "electron:mac": "npm run build:prod && electron-builder build --mac",
-    "test": "ng test",
-    "e2e": "npm run build:prod && cross-env TS_NODE_PROJECT='e2e/tsconfig.e2e.json' mocha --timeout 300000 --require ts-node/register e2e/**/*.e2e.ts",
-    "version": "conventional-changelog -i CHANGELOG.md -s -r 0 && git add CHANGELOG.md",
-    "lint": "ng lint"
-  },
-  "devDependencies": {
-    "@angular-builders/custom-webpack": "8.4.1",
-    "@angular-devkit/build-angular": "0.901.10",
-    "@angular-eslint/builder": "0.0.1-alpha.18",
-    "@angular/cli": "9.0.3",
-    "@angular/common": "9.0.2",
-    "@angular/compiler": "9.0.2",
-    "@angular/compiler-cli": "9.0.2",
-    "@angular/core": "9.0.2",
-    "@angular/forms": "9.0.2",
-    "@angular/language-service": "9.0.2",
-    "@angular/platform-browser": "9.0.2",
-    "@angular/platform-browser-dynamic": "9.0.2",
-    "@angular/router": "9.0.2",
-    "@ngx-translate/core": "12.1.2",
-    "@ngx-translate/http-loader": "4.0.0",
-    "@types/jasmine": "3.5.6",
-    "@types/jasminewd2": "2.0.8",
-    "@types/mocha": "7.0.1",
-    "@types/node": "12.11.1",
-    "@typescript-eslint/eslint-plugin": "2.20.0",
-    "@typescript-eslint/parser": "2.20.0",
-    "chai": "4.2.0",
-    "codelyzer": "5.1.2",
-    "conventional-changelog-cli": "2.0.34",
-    "core-js": "3.1.4",
-    "cross-env": "7.0.0",
-    "electron": "8.2.4",
-    "electron-builder": "22.3.2",
-    "electron-reload": "1.5.0",
-    "eslint": "6.8.0",
-    "eslint-plugin-import": "2.20.1",
-    "jasmine-core": "3.5.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.4.1",
-    "karma-coverage-istanbul-reporter": "2.1.1",
-    "karma-electron": "6.3.0",
-    "karma-jasmine": "3.1.1",
-    "karma-jasmine-html-reporter": "1.5.2",
-    "mocha": "7.1.1",
-    "npm-run-all": "4.1.5",
-    "rxjs": "6.5.4",
-    "spectron": "10.0.1",
-    "ts-node": "8.3.0",
-    "tslib": "1.10.0",
-    "typescript": "3.7.5",
-    "wait-on": "4.0.0",
-    "webdriver-manager": "12.1.7",
-    "zone.js": "0.10.2"
-  },
-  "engines": {
-    "node": ">=10.13.0"
-  },
-  "dependencies": {
-    "@angular/animations": "9.1.0",
-    "@trodi/electron-splashscreen": "1.0.0",
-    "cloudscraper": "4.6.0",
-    "electron-updater": "4.2.5",
-    "fs-extra": "9.0.0",
-    "jquery": "3.5.1",
-    "mime": "2.4.4",
-    "nconf": "0.10.0",
-    "nodepub": "2.2.0",
-    "request": "2.88.2",
-    "rimraf": "3.0.2",
-    "shelljs": "0.8.3"
-  }
+	"name": "novelscraper",
+	"version": "2.0.0",
+	"description": "App for downloading novels from pirate sites.",
+	"homepage": "https://github.com/HanaDigital/NovelScraper",
+	"author": {
+		"name": "dr_nyt",
+		"email": "shehryar.dev@gmail.com"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/HanaDigital/NovelScraper.git"
+	},
+	"keywords": [
+		"novel",
+		"downloader"
+	],
+	"main": "main.js",
+	"private": true,
+	"scripts": {
+		"postinstall": "electron-builder install-app-deps",
+		"ng": "ng",
+		"start": "npm-run-all -p electron:serve ng:serve",
+		"build": "npm run electron:serve-tsc && ng build --base-href ./",
+		"build:dev": "npm run build -- -c dev",
+		"build:prod": "npm run build -- -c production",
+		"ng:serve": "ng serve",
+		"ng:serve:web": "ng serve -c web -o",
+		"electron:serve-tsc": "tsc -p tsconfig-serve.json",
+		"electron:serve": "wait-on tcp:4200 && npm run electron:serve-tsc && npx electron . --serve",
+		"electron:local": "npm run build:prod && npx electron .",
+		"electron:build": "npm run build:prod && electron-builder build",
+		"electron:deploy": "npm run build:prod && electron-builder build --windows --publish always",
+		"test": "ng test --watch=false",
+		"test:watch": "ng test",
+		"e2e": "npm run build:prod && cross-env TS_NODE_PROJECT='e2e/tsconfig.e2e.json' mocha --timeout 300000 --require ts-node/register e2e/**/*.e2e.ts",
+		"version": "conventional-changelog -i CHANGELOG.md -s -r 0 && git add CHANGELOG.md",
+		"lint": "ng lint"
+	},
+	"devDependencies": {
+		"@angular-builders/custom-webpack": "9.2.0",
+		"@angular-devkit/build-angular": "0.1000.2",
+		"@angular-eslint/builder": "0.0.1-alpha.32",
+		"@angular-eslint/eslint-plugin": "0.0.1-alpha.32",
+		"@angular-eslint/eslint-plugin-template": "0.0.1-alpha.32",
+		"@angular-eslint/template-parser": "0.0.1-alpha.32",
+		"@angular/cli": "10.0.2",
+		"@angular/common": "10.0.3",
+		"@angular/compiler": "10.0.3",
+		"@angular/compiler-cli": "10.0.3",
+		"@angular/core": "10.0.3",
+		"@angular/forms": "10.0.3",
+		"@angular/language-service": "10.0.3",
+		"@angular/platform-browser": "10.0.3",
+		"@angular/platform-browser-dynamic": "10.0.3",
+		"@angular/router": "10.0.3",
+		"@ngx-translate/core": "13.0.0",
+		"@ngx-translate/http-loader": "6.0.0",
+		"@types/jasmine": "3.5.11",
+		"@types/jasminewd2": "2.0.8",
+		"@types/mocha": "7.0.2",
+		"@types/node": "12.12.6",
+		"@typescript-eslint/eslint-plugin": "3.6.0",
+		"@typescript-eslint/eslint-plugin-tslint": "3.6.0",
+		"@typescript-eslint/parser": "3.6.0",
+		"chai": "4.2.0",
+		"conventional-changelog-cli": "2.0.34",
+		"core-js": "3.6.5",
+		"cross-env": "7.0.2",
+		"electron": "^10.1.3",
+		"electron-builder": "22.8.1",
+		"electron-reload": "1.5.0",
+		"eslint": "7.4.0",
+		"eslint-plugin-import": "2.22.0",
+		"jasmine-core": "3.5.0",
+		"jasmine-spec-reporter": "5.0.2",
+		"karma": "5.1.0",
+		"karma-coverage-istanbul-reporter": "3.0.3",
+		"karma-electron": "6.3.0",
+		"karma-jasmine": "3.3.1",
+		"karma-jasmine-html-reporter": "1.5.4",
+		"mocha": "7.2.0",
+		"npm-run-all": "4.1.5",
+		"rxjs": "6.6.0",
+		"spectron": "11.1.0",
+		"ts-node": "8.10.2",
+		"tslib": "2.0.0",
+		"typescript": "3.9.6",
+		"wait-on": "5.0.1",
+		"webdriver-manager": "12.1.7",
+		"zone.js": "0.10.3"
+	},
+	"engines": {
+		"node": ">=10.13.0"
+	},
+	"dependencies": {
+		"@trodi/electron-splashscreen": "1.0.0",
+		"cloudscraper": "4.6.0",
+		"electron-updater": "4.3.5",
+		"mime": "2.4.6",
+		"nconf": "0.10.0",
+		"nodepub": "2.2.0",
+		"request": "2.88.2",
+		"rimraf": "3.0.2",
+		"shelljs": "0.8.4"
+	}
 }
