@@ -1,130 +1,81 @@
 import { Component, NgZone } from '@angular/core';
-import { ElectronService } from './core/services';
-import { TranslateService } from '@ngx-translate/core';
 import { AppConfig } from '../environments/environment';
-import { BrowserWindow } from 'electron';
-import { RouterOutlet } from '@angular/router';
-
-// Import Library Service
-import { LibraryService } from './library.service';
-import { ngModuleJitUrl } from '@angular/compiler';
-
-const {remote} = require('electron');
+import { ipcRenderer, remote } from 'electron';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+	selector: 'app-root',
+	templateUrl: './app.component.html',
+	styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  public menuButtons: any;
-  public menuBackground: string;
-  public buttonHighlight: string;
+	version = "UPDATING";
+	isUpdate = false;
+	isUpdating = false;
 
-  public version: string = "UPDATING";
-  public updating: boolean = false;
+	menuButtons: HTMLCollectionOf<Element> = document.getElementsByClassName("button");
 
-  constructor(
-    public electronService: ElectronService,
-    private translate: TranslateService,
-    private library: LibraryService,
-    private zone: NgZone
-  ) {
-    translate.setDefaultLang('en');
+	constructor(zone: NgZone) {
+		console.log('Production: ', AppConfig.production);
 
-    // Define button colors to maintain
-    this.menuButtons = document.getElementsByClassName("menuButton");
-    this.menuBackground = "#033e63";
-    this.buttonHighlight = "linear-gradient(90deg, rgba(6,113,179,1) 0%, rgba(3,62,99,1) 100%)";
+		// Load Version Number
+		ipcRenderer.send('app_version');
+		ipcRenderer.on('app_version', (event, arg) => {
+			ipcRenderer.removeAllListeners('app_version');
+			console.log(arg.version);
+			zone.run(() => {
+				this.version = 'V' + arg.version;
+			});
+		});
 
-    // console.log('AppConfig', AppConfig);
+		ipcRenderer.on('update_available', () => {
+			ipcRenderer.removeAllListeners('update_available');
+			zone.run(() => {
+				this.isUpdate = true;
+			});
+		});
 
-    if (electronService.isElectron) {
-      // console.log(process.env);
-      // console.log('Mode electron');
-      // console.log('Electron ipcRenderer', electronService.ipcRenderer);
-      // console.log('NodeJS childProcess', electronService.childProcess);
-    } else {
-      // console.log('Mode web');
-    }
+		ipcRenderer.on('update_downloaded', () => {
+			ipcRenderer.removeAllListeners('update_downloaded');
+			zone.run(() => {
+				ipcRenderer.send('restart_app');
+			});
+		});
+	}
 
-    // Load Version Number
-    electronService.ipcRenderer.send('app_version');
-    electronService.ipcRenderer.on('app_version', (event, arg) => {
-      electronService.ipcRenderer.removeAllListeners('app_version');
-      console.log(arg.version);
-      zone.run(() => {
-        this.version = 'V' + arg.version;
-      });
-    });
+	loadPage(id: string): void {
+		for (let i = 0; i < this.menuButtons.length; i++) {
+			this.menuButtons[i].classList.remove("active");
+			if (this.menuButtons[i].id === id) this.menuButtons[i].classList.add("active");
+		}
+	}
 
-    electronService.ipcRenderer.on('update_available', () => {
-      electronService.ipcRenderer.removeAllListeners('update_available');
-      console.log('updating...');
-      zone.run(() => {
-        this.updating = true;
-      });
-    });
+	downloadUpdate(): void {
+		ipcRenderer.send('update-app');
+		this.isUpdating = true;
+	}
 
-    electronService.ipcRenderer.on('update_downloaded', () => {
-      electronService.ipcRenderer.removeAllListeners('update_downloaded');
-      zone.run(() => {
-        this.restartApp();
-      });
-    });
-  }
+	toggleUpdatePrompt(): void {
+		this.isUpdate = false;
+	}
 
-  //UI Control
-  // Higlight the home button and open the home page
-  loadHomePage() {
-      this.deselectButtons();
-      this.menuButtons[0].style.background = this.buttonHighlight;
-  }
+	// Minimize Window
+	minWindow(): void {
+		remote.getCurrentWindow().minimize();
+	}
 
-  // Higlight the sources button and open the sources page
-  loadSourcesPage() {
-      this.deselectButtons();
-      this.menuButtons[1].style.background = this.buttonHighlight;
-  }
+	// Maximize and unMaximize Window
+	maxWindow(): void {
+		remote.getCurrentWindow().isMaximized() ?
+			remote.getCurrentWindow().unmaximize()
+			: remote.getCurrentWindow().maximize();
+	}
 
-  // Higlight the library button and open the library page
-  loadLibraryPage() {
-      this.deselectButtons();
-      this.menuButtons[2].style.background = this.buttonHighlight;
-      this.library.loadNovels();
-  }
+	// Close Window
+	closeWindow(): void {
+		remote.getCurrentWindow().close();
+	}
 
-  // Higlight the setting button and open the library page
-  loadSettingPage() {
-    this.deselectButtons();
-    this.menuButtons[3].style.background = this.buttonHighlight;
-    this.library.loadNovels();
-  }
-
-  // Deselect all menu buttons
-  deselectButtons() {
-      for(let i = 0; i < this.menuButtons.length; i++) {
-        this.menuButtons[i].style.background = this.menuBackground;
-      }
-  }
-
-  // Window Controls
-  // Minimize Window
-  minWindow() {
-    remote.getCurrentWindow().minimize();
-  }
-
-  // Maximize and unMaximize Window
-  maxWindow() {
-    remote.getCurrentWindow().isMaximized() ? remote.getCurrentWindow().unmaximize() : remote.getCurrentWindow().maximize();
-  }
-
-  // Close Window
-  closeWindow() {
-    remote.getCurrentWindow().close();
-  }
-
-  restartApp() {
-    this.electronService.ipcRenderer.send('restart_app');
-  }
+	restartApp(): void {
+		ipcRenderer.send('restart_app');
+	}
 }
